Extract keyword builders to stocks module
